Refactoring further, adding natlas namespace instead of prefixing files with natlas
This also splits the threadscan class and all associated work into it's own file. Main's primary job is to feed work to threads now.
